homework09 test added

+++ HomeWork09/src/main/java/ru/otus/homework09/dao/UsersDAO.java
@@ -13,7 +13,7 @@ public class UsersDAO {
         orm = new ORM(connection);
     }
 
-    public UsersDataSet getUser(int id) {
+    public UsersDataSet getUser(Long id) {
         try {
             return orm.loadObject(UsersDataSet.class, id);
         } catch (Exception e) {
